+++ src/Console/ConsoleApplication.php
@@ -5,6 +5,8 @@ declare(strict_types=1);
 namespace Vortex\Console;
 
 use Vortex\Console\Commands\DbCheckCommand;
+use Vortex\Console\Commands\MakeCommandCommand;
+use Vortex\Console\Commands\MakeMigrationCommand;
 use Vortex\Console\Commands\MigrateCommand;
 use Vortex\Console\Commands\DoctorCommand;
 use Vortex\Console\Commands\MigrateDownCommand;
@@ -36,6 +38,8 @@ final class ConsoleApplication
         $app->register(new DbCheckCommand($basePath));
         $app->register(new MigrateCommand($basePath));
         $app->register(new MigrateDownCommand($basePath));
+        $app->register(new MakeMigrationCommand($basePath));
+        $app->register(new MakeCommandCommand($basePath));
         $app->register(new QueueWorkCommand($basePath));
         $app->register(new QueueFailedCommand($basePath));
         $app->register(new QueueRetryCommand($basePath));